+++ package.json
@@ -9,8 +9,8 @@
 		"dev:admin": "dotenv -- turbo run dev --filter admin",
 		"dev:mail": "dotenv -- pnpm --filter mail run dev:mail",
 		"start": "dotenv -- turbo run start",
-		"lint": "biome lint .",
-		"format": "biome format --write . && biome check --apply-unsafe .",
+		"lint": "biome lint",
+		"format": "biome format --write&& biome check --write --unsafe",
 		"db:push": "dotenv -- pnpm --filter db run push",
 		"db:push:prod": "dotenv -- pnpm --filter db run push:prod",
 		"db:studio": "dotenv -- pnpm --filter db run studio",
