+++ packages/db/src/query/user.ts
@@ -1,6 +1,16 @@
-import { and, desc, eq, gt, lte, or, sql } from "drizzle-orm";
+import { and, desc, eq } from "drizzle-orm";
 import db from "../";
-import { historyReports, historyUserData, reports, reportConfig, session, user, userData, userExperimentData } from "../schema";
+import {
+    historyReports,
+    historyUserData,
+    reportConfig,
+    reports,
+    session,
+    token,
+    user,
+    userData,
+    userExperimentData,
+} from "../schema";
 import type { UserSelectType } from "../types/types";
 import { TokenType } from "../types/types";
 import { getReportConfigByUserId } from "./reports";
