+++ scripts/insert-energy-data.ts
@@ -1,6 +1,6 @@
-import { insertRawEnergyValues } from "@energyleaf/db/query";
 import fs from "node:fs";
 import path from "node:path";
+import { insertRawEnergyValues } from "@energyleaf/db/query";
 
 export async function insertEnergyData(args: string[]) {
     const sensorId = args[0];
@@ -12,12 +12,14 @@ export async function insertEnergyData(args: string[]) {
     const fileContent = fs.readFileSync(path.join(filePath, "demo.json"));
     const day = new Date().getDate() + 1;
     const month = new Date().getMonth();
-    const data = (JSON.parse(fileContent.toString()) as {
-        id: string;
-        sensorId: string;
-        value: number;
-        timestamp: string;
-    }[]).map((d) => {
+    const data = (
+        JSON.parse(fileContent.toString()) as {
+            id: string;
+            sensorId: string;
+            value: number;
+            timestamp: string;
+        }[]
+    ).map((d) => {
         const dataDate = new Date(d.timestamp);
         dataDate.setDate(day);
         dataDate.setMonth(month);
